windowing/egl: add Mali support
Add the class CEGLNativeTypeMali which uses the Mali binary driver
(https://github.com/linux-sunxi/sunxi-mali) to initialize an EGL
context.  It can be used to run Kodi on the Mali frame buffer without
X11.

This was tested successfully on a CubieTruck and a Cubieboard.

The screen resolution is currently hard-coded to the size of my TFT
monitor.  This part will need to be rewritten.

Signed-off-by: memeka <mihailescu2m@gmail.com>