wayland: Get user-data through container_of
There is some dark magic that goes on here. Essentially, container_of
uses offsetof in order to get the struct that contained the listener.

However, it only works with structs, so we have to have an inner
struct for the listener and the this pointer and then an outer
class which contains the inner struct.

Signed-off-by: memeka <mihailescu2m@gmail.com>